pyproject.toml
[project]
name = "rocknblock-test"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "whatislove118",email = "pochikalin@gmail.com"},
]
dependencies = [
    "django>=5.0.2",
    "djangorestframework>=3.14.0",
    "psycopg>=3.1.18",
    "orjson>=3.9.14",
    "dj-database-url>=2.1.0",
    "whitenoise>=6.6.0",
    "django-cors-headers>=4.3.1",
    "drf-orjson-renderer>=1.7.2",
    "drf-spectacular>=0.27.1",
    "shortuuid>=1.0.11",
    "web3>=6.15.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
local = [
    "gunicorn>=21.2.0",
]
dev = [
    "nox>=2023.4.22",
    "mypy>=1.8.0",
    "flake8>=7.0.0",
    "djangorestframework-stubs>=3.14.5",
    "ipython>=8.21.0",
    "autoflake>=2.2.1",
    "django-stubs>=1.16.0",
    "pre-commit>=3.6.1",
    "tox-pdm>=0.7.2",
    "pdm-vscode>=0.1.1",
]


[tool.pdm]
distribution = false
